feat: introduce mqtt module and sqlite-backed commands (master)
- Add MQTT client wrapper with command executor wiring
- Move command handlers and state publisher under mqtt package
- Introduce domain models and sqlite storage for commands
- Add goose-based sqlite migrations and startup migration step
- Extend config with MQTT and storage sections and HTTP shutdown
- Replace cmd-level auth/sqlite wiring with internal packages
- Adjust sqlc config and queries for command params smallint type
- Update .gitignore and dependencies for new storage and tooling

+++ data/migrations/sqlite/20260210133628_create_command_params.sql
@@ -2,9 +2,9 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS command_params
 (
-    command_id TEXT NOT NULL REFERENCES commands (id),
-    name      VARCHAR(255) NOT NULL,
-    type      INTEGER NOT NULL CHECK (type >= 1 AND type <= 3),
+    command_id TEXT         NOT NULL REFERENCES commands (id),
+    name       VARCHAR(255) NOT NULL,
+    type       SMALLINT     NOT NULL CHECK (type >= 1 AND type <= 3),
 
     PRIMARY KEY (command_id, name)
 );
